# Add faster ``TocTree._toctree_copy`` method